:sparkles:[Feat] : 작가별 작품 조회 (#27)
