+++ src/main/java/com/example/artstoryage/service/ArtWorkCommandService.java
@@ -1,7 +1,5 @@
 package com.example.artstoryage.service;
 
-import java.util.List;
-
 import com.example.artstoryage.domain.ArtWork;
 import com.example.artstoryage.domain.member.Member;
 import com.example.artstoryage.dto.request.ArtWorkRequestDto.*;
@@ -11,8 +9,6 @@ public interface ArtWorkCommandService {
   ArtWork regArtWork(Member member, RegArtWorkRequest request);
 
   ArtWork allowArtWork(Long artWorkId);
-
-  List<ArtWork> getArtWorksByArtist(Long artistId);
 
   void deleteArtWork(Long artWorkId);
 }
