+++ src/main/java/com/example/artstoryage/service/ArtWorkQueryService.java
@@ -1,0 +1,9 @@
+package com.example.artstoryage.service;
+
+import java.util.List;
+
+import com.example.artstoryage.domain.ArtWork;
+
+public interface ArtWorkQueryService {
+  List<ArtWork> getArtWorksByArtist(Long artistId);
+}
